requirements.txt
absl-py==2.1.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
einops==0.8.0
filelock==3.13.1
fonttools==4.51.0
fsspec==2024.2.0
grpcio==1.63.0
huggingface-hub==0.23.0
idna==3.7
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
prettytable==3.10.0
protobuf==5.26.1
pycocotools==2.0.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
timm==0.9.16
torch==2.3.0+cu118
torchaudio==2.3.0+cu118
torchseg==0.0.1a4
torchvision==0.18.0+cu118
tqdm==4.66.2
triton==2.3.0
typing_extensions==4.9.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
xml-python==0.4.3
